multiple bug fixes
dash attack hitboxes now deactivate if you cancel dash attack early. same for up b.

also, run can transition to jab on the very tiny window that the skid animation starts during run. (technically "run" as a state doesn't end for a tiny period of time after you let go of the run button. during this time, it's checking to see if you're going to input the opposite direction. if so, it transitions to pivot.)